Use VCR's hooks to simply record requests in git
VCR has pluggable Serializers and Persisters. These sound very handy,
but in reality they're quite tightly coupled to VCR's view of the world,
which expects you to be storing all of your interactions into one file,
on disk. In our case we want a separate file for each interaction, and
we also want to split the html out from the rest of the metadata for
easier diffing.

After spending some time trying to wrestle serializers and persisters
into shape I've now changed tactics and am instead just installing a
null serializer and null logger. This means we can simply use VCR's
hooks to know when there's been an http request and record it.

At some point in the future we're going to need to replay these
requests, but for now this is good enough.